telegram_bot/requirements.txt
python-telegram-bot>=20.0
pyzbar>=0.1.9
Pillow>=9.0.0
python-dotenv>=1.0.0
requests>=2.28.0
